using list_db with size

+++ 02_Using_list_DB/test_List_DB.py
@@ -5,22 +5,28 @@ import hashlib
 
 #바이러스 DB
 VirusDB = {
-    '117599303ed472727a2cbe1343ad6989:EICAR TEST',
-    'ef5a19cc8eed8de1621f6a0591e374e4:Dummy TEST'
+    '86:117599303ed472727a2cbe1343ad6989:EICAR TEST',
+    '80:ef5a19cc8eed8de1621f6a0591e374e4:Dummy TEST'
 
 }
 
 #프로그램 동작 중 실제로 사용하는 DB용 리스트
 vdb= []
+vsize = []
+
 
 #VirusDB --> vdb로 작업 진행
 def MakeVirusDB():
     for pattern in VirusDB:
         t = []
         v = pattern.split(':')
-        t.append(v[0])
         t.append(v[1])
+        t.append(v[2])
         vdb.append(t)
+
+        size = int(v[0])
+        if(vsize.count(size)) == 0:
+           vsize.append(size)
 
 
 #vdb에서 md5 비교
@@ -39,21 +45,28 @@ if __name__ == '__main__':
         exit(0)
 
     fname = sys.argv[1]
-    fp = open(fname, 'rb')
-    fbuf = fp.read()
-    fp.close()
+    size = os.path.getsize(fname)
+
+    if vsize.count(size):
+           
+        fp = open(fname, 'rb')
+        fbuf = fp.read()
+        fp.close()
 
 
-    m = hashlib.md5()
-    m.update(fbuf)
-    fmd5 = m.hexdigest()
+        m = hashlib.md5()
+        m.update(fbuf)
+        fmd5 = m.hexdigest()
 
-    ret, vname = SearchDB(fmd5)
-# vdb에 존재하는 MD5 비교
-    if ret == True:
-        print 'virus'
-        print '%s : %s' %(fname, vname)
-        os.remove(fname)
+        ret, vname = SearchDB(fmd5)
+    # vdb에 존재하는 MD5 비교
+        if ret == True:
+            print 'virus'
+            print '%s : %s' %(fname, vname)
+            os.remove(fname)
+        else:
+            print 'No virus'
+            print '%s : ok' %(fname)
     else:
-        print 'No virus'
-        print '%s : ok' %(fname)
+           print 'No Virus'
+           print '%s : ok' %(fname)
